Hex Encrypt
Added in functionality where the user's email address and passwords are saved to the UsrCrdn.txt file in hex format instead of standard ASCII, so that the file cannot be quickly opened to see a user's password by someone else. Allowing for the user's information to be safely stored in the system for quick log in.
TO DO: Add in a hex decryption to allow for quick login on re-opening the program.